@@ -1,22 +1,20 @@
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace RedDog.AccountingModel
+namespace RedDog.AccountingModel;
+
+[Table(nameof(Customer))]
+public class Customer
 {
-    [Table(nameof(Customer))]
-    public class Customer
-    {
-        [Column(TypeName = "nvarchar(36)")]
-        [Key]
-        public string LoyaltyId { get; set; }
+    [Column(TypeName = "nvarchar(36)")]
+    [Key]
+    public required string LoyaltyId { get; set; }
 
-        [Column(TypeName = "nvarchar(50)")]
-        [Required]
-        public string FirstName { get; set; }
+    [Column(TypeName = "nvarchar(50)")]
+    [Required]
+    public required string FirstName { get; set; }
 
-        [Column(TypeName = "nvarchar(50)")]
-        [Required]
-        public string LastName { get; set; }
-
-    }
+    [Column(TypeName = "nvarchar(50)")]
+    [Required]
+    public required string LastName { get; set; }
 }
