@@ -1,22 +1,21 @@
 using System.Text.Json.Serialization;
 
-namespace RedDog.ReceiptGenerationService.Models
+namespace RedDog.ReceiptGenerationService.Models;
+
+public class OrderItemSummary
 {
-    public class OrderItemSummary
-    {
-        [JsonPropertyName("productId")]        
-        public int ProductId { get; set; }
-        
-        [JsonPropertyName("productName")]
-        public string ProductName { get; set; }
-        
-        [JsonPropertyName("quantity")]
-        public int Quantity { get; set; }
+    [JsonPropertyName("productId")]
+    public int ProductId { get; set; }
 
-        [JsonPropertyName("unitCost")]
-        public decimal UnitCost { get; set; }
+    [JsonPropertyName("productName")]
+    public string ProductName { get; set; } = string.Empty;
 
-        [JsonPropertyName("unitPrice")]
-        public decimal UnitPrice { get; set; }
-    }
+    [JsonPropertyName("quantity")]
+    public int Quantity { get; set; }
+
+    [JsonPropertyName("unitCost")]
+    public decimal UnitCost { get; set; }
+
+    [JsonPropertyName("unitPrice")]
+    public decimal UnitPrice { get; set; }
 }
