@@ -1,36 +1,33 @@
-using System;
-using System.Collections.Generic;
 using System.Text.Json.Serialization;
 
-namespace RedDog.ReceiptGenerationService.Models
+namespace RedDog.ReceiptGenerationService.Models;
+
+public class OrderSummary
 {
-    public class OrderSummary
-    {
-        [JsonPropertyName("orderId")]
-        public Guid OrderId { get; set; }
+    [JsonPropertyName("orderId")]
+    public Guid OrderId { get; set; }
 
-        [JsonPropertyName("orderDate")]
-        public DateTime OrderDate { get; set; }
+    [JsonPropertyName("orderDate")]
+    public DateTime OrderDate { get; set; }
 
-        [JsonPropertyName("orderCompletedDate")]
-        public DateTime? OrderCompletedDate { get; set; }
+    [JsonPropertyName("orderCompletedDate")]
+    public DateTime? OrderCompletedDate { get; set; }
 
-        [JsonPropertyName("storeId")]
-        public string StoreId { get; set; }
+    [JsonPropertyName("storeId")]
+    public string StoreId { get; set; } = string.Empty;
 
-        [JsonPropertyName("firstName")]
-        public string FirstName { get; set; }
+    [JsonPropertyName("firstName")]
+    public string FirstName { get; set; } = string.Empty;
 
-        [JsonPropertyName("lastName")]
-        public string LastName { get; set; }
+    [JsonPropertyName("lastName")]
+    public string LastName { get; set; } = string.Empty;
 
-        [JsonPropertyName("loyaltyId")]
-        public string LoyaltyId { get; set; }
+    [JsonPropertyName("loyaltyId")]
+    public string LoyaltyId { get; set; } = string.Empty;
 
-        [JsonPropertyName("orderItems")]
-        public List<OrderItemSummary> OrderItems { get; set; }
+    [JsonPropertyName("orderItems")]
+    public List<OrderItemSummary> OrderItems { get; set; } = new();
 
-        [JsonPropertyName("orderTotal")]
-        public decimal OrderTotal { get; set; }
-    }
+    [JsonPropertyName("orderTotal")]
+    public decimal OrderTotal { get; set; }
 }
